Model file content availability and access methods
This includes a significant change of concept. Rather than modeling
storage systems in an abstract way, we model access methods. This
makes things a lot leaner and enable lean connections to external
implementations. For example, any and all git annex special remotes
are cover by a single method class.

The example also includes a nesting of access methods: an extraction
from an archive that can be downloaded via HTTP GET.

The notion of credential requirements could also be worked into this
fabric in the future.