Remove `identifying` ontology components
The implications of declaring an identifier are to complex to justify
such a simplification. At least at this point. See for example:
https://github.com/psychoinformatics-de/datalad-concepts/issues/17

Let's stick to declaring identifiers in the scope of a particular
schema, rather than ontology-wide. Depending on the data structure
(not semantics), some schemas might get away without using any (global)
identifiers, and rely exclusively on nesting.