Establish `meta_utils` ontology component
Use a consistent "meta" prefix for slots and mixin. The purpose of these
utilities is identification and type inference of METADATA objects.

This is an important distinction. Whether or not an entity has an
identifier, we still need an identifier for the corresponding metadata
document about it. With this change, we choose to align the respective
property with naming and format to the type designator for metadata
objects.