Git provenance update
Settled on a provisional approach to record author separate from a
committer. It is done, by modeling both aspects as activities
(Authoring, Committing) that can have the PROV-typical associations
with Agents, and the authoring activity informing the committing
activity.

This may not be the simplest model (one could also create a flatter
structure with properties that have more complex semantics. However,
this is more in the spirit of PROV, and I suspect that it would work
better for a generic model (vs. a more "front-end" schema).

Time will tell.