New `git` ontology component with new `GitTracked` mixin
This aims to be the foundation for data models of Git-tracked content.
The availability of a globally unique identifier in this context
allows for great flexibility in model specification. At the same time,
introducing a `identifier:true` property in a class hierarchy can be
problematic, because turning an existing base-class property into an
identifier in a derived class does not seem to be possible
(https://github.com/psychoinformatics-de/datalad-concepts/issues/30).

This approach uses a dedicated mixin class for this purpose.

The identifier itself uses the `meta_id` slot, and is intentionally
distinct from the `gitsha` property. The identifier is composed from
the `gitsha` by adding a corresponding namespace prefix. This keeps
the identifier property uniform across unrelated classes, while
also preserving the semantics of the identifier type.