Add schema for an "object stream" dataset description
This is conceptually already quite close to what we need to get at --
for serializing datasets, starting from their Git repository
representation.

This is only a minimal sketch at this point, but it demos how the
recently found duality of semantic-defining and structure-defining
classes would look like in this contex.

There are a bunch of open question that need more consideration.
Something like:

- Should we make the SE-level classes define `exact_mappings` to their
  ontology counterparts, or use discovery like shown in
  https://github.com/psychoinformatics-de/datalad-concepts/issues/37

- what kind of value do we use for `meta_type`? CURIEs? or plain class
  names?

- ...