Compose "full" example for components schema
See src/examples/datalad-dataset-components/ContainerSE-DatasetWFiles.yaml

It works in principle, but I was not yet able to get the description of
an AnnexedFileSE's `distribution` to work (see comments inside).

This is strange, because this part is more or less taken verbatim from
the `datalad-dataset-version` schema, and
`src/examples/datalad-dataset-version/DataladDatasetVersionSE-full.yaml`
shows it to be working properly.

As elaborated on inside, the JSON schema code generator does not include
a class that is required for validation. However, it does reference it.