Remove `datalad-dataset-version` schema
This has outlived its purpose. A similar schema needs to be reborn
for the purpose of describing non-DataLad dataset. In order to avoid the
need for index-related identifiers, it needs to be fully
nested/monolithic.

However, this needs more thought and this hybrid causes more maintenance
work than it helps at the moment.