Switch to linked-trees for describing a dataset distribution (DataLad)
The established `qualified_part` pattern is no applied to items in a
directory. Given that only immediate children are linked, there is
no need for a more complex `at_location` property with a `PosixRelPath`
type anymore.

`FilesystemDirectoryItem` is now directly modeled as a `QualifiedPart`.
This further simplifies and homogenizes the model.