Add example showing dataset nesting
The key aspect is that a `DataladDatasetVersion` is linked to another
`DataladDatasetVersion` via `has_part`. The specific "mountpoint is,
however, not recorded in this relationship.

This information is captured by linking the subdataset version tree
directly in the superdataset version tree.

This enables (IMHO) relatively straightforward use case. Any consumer
that requires the exact file tree of a superdataset (including all
linked subdatasets) can simply ignore the `has_part` relationship.

When subtrees have to be associated with particular subdatasets, a
consumer can either start traversing from the respective subdatasets
directly, a cross-reference subdatasets `tree` properties when
traversing the superdataset tree.

In a multiversion, multidataset record, there is no duplication of tree
information resulting from this organization. Any identical subtrees
(shared across datasets (and/or versions thereof) would be deduplicated
automatically.