Merge `AnnexedFile|AnnexDistribution` -> `StableAnnexKey`
This move simplifies the class set, and also the data structure of the
`datalad-dataset-components` (see example inside).

A `StableAnnexKey` is now a `Distribution`, and the corresponding
`Resource` subclass is dropped without a replacement. There seems to be
no loss of expressiveness.

This makes `StableAnnexKey` the only current "content" alternative to
a `GitBlob` (which has been renamed from `FileInGit` to emphasize the
focus on the content, rather than any given name).

A dedicated `annex-key` namespace has been introduced too

Closes #50