Introduce document-local object identification by a code
This implements the approach identified by @jsheunis in #66.

Specifically, the combination of a new type `Code` and a new
`schema_utils` slot `meta_code` enable the identification and
referencing of metadata objects without the need for a globally
unique identifier (see the documentation of the respective
ontology additions).

The new capability is demo'd on `GitUserAgent` description
in the `datalad-dataset-components` schema. Here the email
is used a document-constraint identification code. A new
`multiversion-dataset` example has been added to showcase a
complete example of this.

In the `datalad-dataset-components` schema itself, the key `ComponentSE`
class has been subclassed twice for objects with a full URI identifier
and for an identifier code. This was both object types can be used in
the same metadata documents. Moreover, the class derivations clearly
show which identification schema is used for which object type.
This approach likely needs to be generalized further and eventually
moved into `schema_utils`.

Closes #66