Provide a definition for any `*Object` class in the ontology
The motivation here is to have the `dlco:...` `class_uri` used in
individual schemas resolve to an actual documentation page.

Even though the class definitions in individual schemas will be somewhat
different, these classes capture the essence of the concepts the
represent in a concrete metadata document.

This ontology component is also useful for starting a new schema. It
will always represent the full set of possible metadata object types
(that can be specialized for a schema). As such, this file is a king of
template for a new schema.