Simplified class names for qualified relation pattern
The object representing a specific qualified relation typically has
another semantic aspect that is sufficient for distinguishing from other
concept, with no need to carry a `Qualified` prefix around.

For example an object representing a `Part` relationship can simply say
what is relevant for that specific relationship, e.g. `GitShaIDedPart`.
It does not add any information to call it `QualifiedGitShaIDedPart`.