Support license declarations for resources and distributions
This is cutting quite a few corners. However, as the examples show this
allows for the specification of standard licenses (by identifier), and
also the inclusion of arbitrary license terms as full text.

Few choices have been made:

There can only be one license per entity. Cases of dual-licenses,
exception terms etc need to be specified with a custom LicenseDocument.
SPDX has the vocab for that.

DatasetVersionObject has received a custom_licenses property. It has no
other purpose than to be a container for custom license declarations.
So it is purely needed for schema validation, hence has no counter-part
in the `DatasetVersion` concept class.