Establish pattern to express organizational affiliations
The key idea is to capture any organization via a
`qualified_attribution` in `DatasetVersion`. Such a specification uses
a dedicated `Organization` class and a `meta_code` property to
yield a document-local identifier that can be used to reference the
organization.

The `qualified_attribution` also documents the role of an organization.
When it had no particular role and is only included in order to express
an affiliation, the `marcrel:sht` role (Supporting Host) appears to
be a good description (hosting the person, or the person's work).

A dedicated example is included.

The Organization class itself remains rather bare for now.