Basic pattern to describe related publications
This employs the same basic principle of a `qualified_relation`, as for
other related entity. A publication is describes together with its roles
regarding the relationship with the dataset.

A range of useful relationship roles are provided by the CiTO ontology,
two are included here. `cites_as_authority` and
`is_cited_as_data_source_by` can be used jointly to identify a primary
publication on a dataset (i.e., a data descriptor).

The `Publication` class is a bare-bone placeholder for now.