Align entity role specification with agent role specification
They are conceptually the same thing, so we better use the same
approach.

The previously envisioned approach via an enum is too constricting. We
will never have an exhaustive description of roles that could be
relevant.

So in this new way, any CURIE can be used to define a role. This is the
most flexible, yet uncompromisingly precise way to declare roles.

Admittedly, it also opens the door to needless heterogeneity and
complexity, especially compared to an enum aproach, which would not even
require a prefix.

We have to commit to providing simple and clear example that cover the
most common use cases, so nobody needs to get creative for no reason.