Support dataset identifier specification
There is no dedicated property for DOIs. This could be singled out, and
declared more important that any other. However, in practice there are
tons of ePIDs that all need a place to go.

We may enhance the setup with an `IdentifierObject` class in the future
which would be able to associate an identifier type with the actual
identifier string, i.e. a `qualified_identifier` to apply our standard
pattern again.

Ping #89