Rough sketch to specify the distribution of a `DatasetVersionObject`
The idea is to stay structurally close to the qualified part pattern
used for the `datalad-dataset-components` schema, but not require any
global identifiers.

The class names reflect the broad applicability of this specification.
`FileContainerObject` is a logical structure to name/locate arbitrary
digital content (`DigitalDocumentObject`). There is no assumption that
a `DigitalDocument` is a physical file at a providing data service. it
could very well be a blob in a database or the outcome of a query.