Enable `qualified_attribution` for `Distribution`s
This made it necessary to move the inline definition of agents. This is
now down at the `DatasetVersionObject` level via `was_attributed_to`.

This is essentially a list of all relevant agents without any further
qualification re the nature of the attribution. This is then done via
`qualified_attribution` of whatever object via its `meta_code`
identifier.