Enhance enough to capture the penguins example
In particular settle on a system to capture arbitrary identifiers with
sufficient clarity, without having to introduce a dedicated property
for each of them. Shown here on the examples of ORCID, ROR, and DOI.

This example now also shows better how it is possible to mix and match
local and global identifiers.