Distinguish `type` and `meta_type`
The former is `RDF:type` whereas the latter is a linkml-specific
construct that enables the tooling to pick the right schema class
for validation and data structure forming.

The changeset include an immediate application: We want to be able to
state that an `Entity` is a peer reviewed journal, while not requiring
a dedicated class, because this statement has no impact on the metadata
structure wrt to the defining schema.