First rough mapping of a DataLad dataset export onto `data-distribution`
I think this worked pretty good for a first attempt!

A number of examples have been added that roughly reflect the
compartmentalization of metadata record for storage in a datalad dataset

- annex key metadata -> Distribution
- git blob metadata -> Distribution
- git tree metadata -> Distribution
- git commit -> Resource
  - this is also the anchor for provenance records
  - here demo'd for a derivation from a parent commit via an unspecific
    "change"
- datalad dataset (version-less) -> Resource

Structurally, `EntityInfluence` has been enhanced to accept multiple
entities. This makes specifying a homgeneous set of roles for a number
of entities more space-efficient. It also enables the annotation of
merge commits (that have multiple parent commits).

This should make the `datalad-dataset-component` schema obsolete.