Data access description example
This is using a fictitious installation of the RWTH RDM system Coscine
as an example. Most needed aspects can be mapped onto properties that
are provided by DCAT. Missing pieces are:

1. Specification of full details of how to perform an actual download request
2. Required authentication

Specifically `DCAT:endpointURL` is just "root location or primary
endpoint of the service", but most APIs offer more than a download API,
hence need some further specification. Coscine uses a particular URL
path for its "get blob" API, which needs to be parameterized with three
parameters. These parameters needs to be put into a qualified-relation
(Distribution-DataService).

The approach via `endpoint_path` and `access_parameter` is arbitrary. I
could not find a common pattern or vocabulary to do this. An equallly
good (from my POV) alternative would be:

- `DataService.download_url_template` -- with some convention to put
  placeholders into a download URL
- `QualifiedAccess.download_url_parameters` with a list of name/value
  pairs

Yet another alternative would be to encode all three shown parameters
into a single "identifier", and call that `access_id` (matching watch
the annex access example is doing). The disadvantage for that would be
that the knowledge of the individual parameters that make up the
compound identifier is lost. If preserved, however, a single-point-fix
of `endpoint_path` could fix up data access in case of an API version
update that does not change the meaning of the identifiers.

Theoretically, we would also need to add information on authentication,
but this is also a vast space of possibilities. A system implementing a
download based on this schema is given a `contact_point`, and also a
dataservice description to be able to present a user with a meaningful
request for entering a credential, at least.