Distingiush `Characteristic` from `Property`
The (more or less arbitrary) definition is

- a property is something that is observed or measured, something that
  is a fact, given a concrete thing
- a characteristic is the underlying, more generic concept

This is done to be able to distinguish a `Parameter` from a `Property`
later on (or in another schema). A `Parameter` being a more
causal/functional concept, something that is variable and impacts
a system or function.

The reason to base both concepts on the same foundation is the
similarity with respect to the schema design principles. We need to be
able to express parameters without prescribing a particular vocabulary.
Employing the same approach twice in identical fashion (only changing
the semantics) avoids needless complexity.