More proper data-access description support
Major changes are:

- more useful and accurate `distribution` schema description
- more complete `DataService` class
- new `Parameter` concept
- examples that document the capabilities

While most changes more-or-less reflect the continued adoption of `DCAT`
concepts and properties, the introduction of `Parameter` is noteworthy.

`Parameter` is a variant of `Property` and serves a similar purpose
(declare arbitrary additional aspects without prescribing a vocabulary
to do so) with only a change in semantics of the class itself. In
contrast to `Property` (observed or measured, fixed), `Parameter` is a
variable with impact on a system or function.

Closes #171

Via the property `has_parameter` particular parameters can be declared
as supported/needed (e.g., `DataService`), or provided
(`QualifiedAccess`). Examples are included.

`QualifiedAccess` is no longer derived from `EntityInfluence` -- it has
been too much of a stretch. It is now focused on access, and no longer
requires a role specification.

Closes #156