Update of the `distribution` schema
This is primarily in response to the `thing` revision.

Additionally, the `Parameter` concept and the conflict of `AttributeSpecification`
with its required `predicate` vs `has_parameter` needs another pass.
The support for it has been removed here, and needs to be reintroduced
once a suitable solution is available.

This pending update should happen with awareness of
https://github.com/psychoinformatics-de/datalad-concepts/issues/172