Disable of `sdd` example
These are outdated and require an overhaul. However, this requires a
reintroduction of a number of slot definitions that have been removed
from `Thing`. Overall, it would be beneficial to rethink the purpose of
`sdd`. It may be best seen as an example schema that could be used to
craft concrete implementations along its patterns.