Factor out a `roles` schema
This takes the most essential classes and slots from the `prov` schema,
and factors them out. It does so, trying to consolidate the concepts
of PROV's `Influence` and DCAT's `QualifiedRelation` into a single
schema element (`Relationship`).

`Relationship` links two (and only two) things. The previous exception
of a multi-object relationship within `Derivation` is given up for
homogeneity.

A `Relationship` can be characterized by multiple roles. The
corresponding slot `had_roles` is declared to be multivalued (and uses
the plural form in its slot name).

A number of `prov-schema` elements have been removed in an effort to
simplify the resulting data structures:

The slots `agent` and `entity` are replaced by a reuse of thing's
`object` uniformly. The respective range definitions have been moved to
the corresponding association classes.

`Influence` is removed, and `Relationship` is used as a base for more
specialized PROV relationships.