Remove `DOI` features not supported by the SHACL exporter
We need SHACL as a key part of the downstream pipeline. Even if it is
sad to not have `creator` and `schema_agency` autofill for a `DOI`, we
have no choice but to disable them until support for `ifabsent` is
added.

Reported at https://github.com/linkml/linkml/issues/2464#issuecomment-2586513489