Switch to inlining as dict for `relations`
Any object under `relations` is a `Things`, and all of them
have a required `id`. So it will always be possible to do this.

This choice is not without concerns. The issue linked below has
some thoughts. In the end, not using lists seems to have more
advantages,
in particular a simpler, direct access by key -- rather than a forced
loop to match on `id`.

Closes: https://github.com/psychoinformatics-de/datalad-concepts/issues/148