Perform tests and conversion in a more self-contained manner
Previously we needed a complicated dance for updating things on the
website incrementally.

See https://github.com/psychoinformatics-de/datalad-concepts/issues/195

This approach was flawed and could not handle breaking transitions.

Here we switch to an approach that alters the schema imports
(temporarily). This achieves a test condition that ensure internal
consistency of all schemas, as they currently are in any specific
commit.