add sh:path annotation to schema_type slot
This addition, together with a patch to linkml's shaclgen.py,
prevents the generation of shacl 'sh:PropertyShape's with
duplicate 'sh:path' values 'rdf:type', because of the fact that
both slots 'type' and 'schema_type' map to 'rdf:type'. This
duplication prevents applications like shacl-vue to distiguish
effectively between the two slots/fields, which means that they
couldn't be used for distinct purposes by users.