Declare `schema_type` range to be `NodeUriOrCurie`
See https://github.com/linkml/linkml/issues/2506 for some background,
and the source of this fix.

This changes enables successful "round-tripping" of data records from
a structured representation (like YAML) to RDF and back to a structure
representation.

Without this change, the conversion from YAML would not recognize
types as schema classes.