Adopt `mappings` slots from linkml
This is one step towards rectifying the ill-conceived `type` slot.
See https://github.com/psychoinformatics-de/datalad-concepts/issues/223
for more information.

Besides the technically issues, the new set of slots offer more
flexibility for users to describe concepts that are not explicitly and
directly models in any given schema -- in the exact same way that the
schemas use to declare alignment of their classes with external
terminologies.