Represent a checksum as a computed identifier
This changes removes a few checksum specific slots.

The `identifiers` schema is diversified to distinguish "issued" from
"computed" identifiers. A difference between the two is that one is
created by an organization-agent, whereas the other is created by a
software-agent.

A `Checksum` is maintained as a dedicated class, but moved from
`distribution`.

With the introduction of multiple identifier subclasses, `Identifier`
itself gained a `type` designator slot that enables the declaration of
type for individual identifiers assigned to `identifiers`.