Align `qualified_relations` with `relations`
And use inlining (of `Relationship` instances) by key.
The difference is that `Relationship` is not `Thing` and does not
provide `id`. However, it has `object` which we can declare to
be a key.

And this is what this change does.

This implies that there can only be a single instance of `Relationship`
for any given `object` within a `qualified_relations` slot.

I think that is OK, and it enables fast, dict-key based access.