Add `annotations` slot to any `Thing`
There is a need to express metadata on a metadata record. For example,
how a metadata record is best presented.

This serves the same purpose as linkml's `annotations` slot, and the use
of the shacl `sh:name` annotation for a class/slot name. However, this
approach also extends into the realm of individual instances.

Use case: Role instances need to be presented in a meaningfully compact
way (also Instruments, etc). `skos:prefLabel` for
https://id.loc.gov/vocabulary/relators/spn.html could simply be set to
`Sponsor` and would offer the information to do this.

LinkML's `annotation_value` slot allows for `linkml:Any` values.
However, this is giving me pain. I decided to not investigate this now
(more important things to do), and leave it at `string` for now.
It should be non-invasive to relax this later on.

An example for a `Role` annotation is included.