Add slot for `skos:prefLabel` -> `preferred_label`
This serves the same purpose as the shacl `sh:name` annotation for a
class/slot name. However, this approach also extends into the realm of
individual instances.

Use case: Role instances need to be presented in a meaningfully compact
way (also Instruments, etc). `skos:prefLabel` for
https://id.loc.gov/vocabulary/relators/spn.html could simply be set to
`Sponsor` and would offer the information to do this.

This is defined as something separate from `name`, `title` etc, to
discourage abusing one purpose for the respective other.