allow strings as `Thing.relations`-values
This commit adds the type "string" as possible
value for an entry in the `relations`-slot of
`Thing`. This allows to represent pure links
to "related" objects by storing a key-value
pair where the key is the `id` of a related
record and the value is ''.

This allows for a much nicer representation
of extracted records (see issue
<https://github.com/psychoinformatics-de/datalad-concepts/issues/275>)