Rename essential `id` slot to `pid`
The name of the slot did not adequately reflect the key property it must
fulfill: a *persistent* identifier. The schema collection and associated
tool implementations are built on this particular property.

It is not standard or to be expected that a metadata record identifier
must be persistent, hence it is useful to reflect this in the name of
slot. It also helps to disambiguate `id` or `pid` from the JSON-LD
built-in `@id`.

Closes: https://github.com/psychoinformatics-de/datalad-concepts/issues/241