Minimal versioning support for `Resource`
This is specifically targeting records that may be archived. So there is
support for a previous version, but not for declaring a next version --
which would not be known at the time of creating a particular version.

I used `version_label`, not plain `version`, to leave space for a
possibly more sophisticated `Version` class.