Add concept of a `Distribution` as/of a `Resource`
This is the common class that distinguished the conceptual resources
(Dataset, Document, etc) from their physical or electronic embodyments.
For example, a conceptual license document from a electronic file
containing the text of that license.

This comes with the slots to relate resources to distributions and a
distribution to respective resource.