More flexible means to express `Thing` relationships
The most important change here is that `Relationship` now supports
the `schema_type` type designator slot. This enables declaring
particular and different relationship types via `qualified_relations`.

This made it necessary to switch `qualified_relations` from a key-based
inlining setup to inlining as a list, because now multiple different
relationship types could be possible between the same two `Thing`
instances.

As part of this effort, the range of the `Relationship` (and
`Statement`) slot `object` has been change to `Thing` (from
`uriorcurie`). This involved no declaring a range in the actual `slot`
definition -- seemingly working around a linkml limitation.

The possibility to express different relationship types via
`qualified_relations` made it possible to remove the slots
`indexed_parts` and `indexed_part_of` from `Resource`, by making
`IndexedResourcePart` and `IndexedResourcePartOf` proper subclasses of
`Relationship`. The, now redundant, `resource` slot was removed, and the
range of `object` adjusted to be `Resource` in these subclasses to
maintain the specificity of validation.

A `TransientRelationship` class was added to the `temporal` schema
component. This can be used to express temporary or time limited
relationship, such as a group membership for a particular time window.

This appears to be a more broadly applicable approach than defining
individual concepts (e.g., `Membership`) that are themselves modeled
to be temporally constrained.

Multiplexing all these semantics onto `qualified_relations` makes the
respective expressions a bit more verbose (in comparison to the very
compact key-value expression that was possible with `indexed_parts`.
But the counter-force to further proliferation of means of expressing
the same things appears to be overall positive.

Closes: https://github.com/psychoinformatics-de/datalad-concepts/issues/288