Introduce schema mixins
The idea is to factor out common parts (i.e. properties) from schema
components that are useful for reuse in schemas that must not
inherit the complex class hierarchy of the main schema components.

One use case are the FLAT schemas. They benefit from using the
same property and type semantics, but cannot inherit the full complexity
of the class hierarchy.

Over time, this might resolve the `properties` schema. It was already
more or less a "mixin".

Mixin schemas should only import from other mixins.