Introduce `Organization` in flat-social
This is the first example of an attempt to maximize orthogonality
of the flat schemas. Only the plain slot `at_location` is taken from
`spatial-mixin`. No range for it is defined there. This is only
done in `flat-social`. But here again, no real "spatial" range
is declared. Instead only `Thing` is used.

The idea is that consuming application schemas will want to flexibly
use specific types (say `City`). Trying to define all possible type
of this kind would blow things up needlessly. `Thing` works, because
everything is a `Thing` in the end, and a GEO URL (`geo:...`) is a
sufficient PID for a `Thing` also.