New `prov-mixin`
This includes a workaround for #316: import things/v1 and declare default range

Issue #316 describes a problem with undefined ranges in mixin slots:
if a schema importing the mixin does not make use of all its slots
(where it would define a range in slot_usage) the OWL conversion
fails for the unused slots.
As a workaround, the slots in the mixin now define a range of Thing.
For completeness, although not necessary to pass the tests, we import
things/v1 (any class that uses the mixin already imports things).