Replace `subject_type` with a generic `is_a` relationship property
This is defined as `dcterms:type`. However, `type` is not used as
the property name to avoid confusion with `rdf:type`, which this is
not. `is_a` is intended to be a non-schema, but semantic classifier.

Remove `SubjectType` also. If an application needs to group particular
types for convenient selection, this can be easily done in that context,
and doesn't need foundational support.