Update linkml shaclgen patch with --include-stripped-prefixes arg
This new option allows inclusion of prefixes that are unused and hence
stripped during rdflib serialization. This makes it possible to specify
namespaces in the prefixes section of a schema that should propagate
to the exported SHACL for downstream use. For SHACL generation purposes
such prefixes need not be included in the emit_prefixes section of a schema