Merge `AssessmentType` and `DistributionType` -> `DataType`
While there is a difference between a data type (image, table) and a
data format (PNG, CSV), there difference is less relevant here.
`DataType` is no more than a class with the purpose to be a type
for individuals that form the collection of relevant types and formats
in a particular scenario.