Approach to multi-role attribution
Not having the qualified relation pattern for this in the flat schemas was a big thing.
However, I dont think that this scales. Even if we ignore the complicated relationships
of people to datasets and just take a publication for now, this is too complex to do
with individual properties: authors, corresponding authors, first authors, senior authors, ...

So this is going for an association class, and specifically attribution. The reason for
that is that the ability to tune this specifically for agent relations, and not having
to deal with the whole scope for a single class.

Not sure if there is anything that a general schema building block could
facilitate here. Hence doing this in the demo context directly.