Remove `Factor` and `Dimension`
This distinction was found to not scale. One an the same thing can
be a factor and a dimension at the same time. If that is true, the
cannot be instanced of different classes under the same PID.

The semantic distinction can be made via thr respective properties
`factors` and `dimensions`.

In the context of the research assets demonstrator these have been
replaced by a merged `XYZConcept` class that serves as the `range`
type for `factors` and `dimensions`.