Introduce classes 'XYZCommunication', 'XYZDocumentExchange', and 'informed_by' slot
The goal here is to be able to model the fact that a project activity is informed by
the act of applying for funding, via some form of document exhange.

'XYZCommunication' is introduced as a class inheriting from 'dlprovmx:Communication',
which has the new 'entity' slot that is meant to be the 'Thing' that is shared between
the two activities via the communication, for example a document.

In order to support document upload in this context. The 'XYZCommunication' is subclassed
to create 'XYZDocumentExchange', where the entity has an updated ranhe 'XYZCommunication'
in addition to receiving the gitAnnexUpload annotation to allow document upload on the
UI side.

In addition, a simple example and related validation config are added.