Experimental approach to make `influenced_by` more usable
In general, this property is difficult from a UI perspective, because
the range of and Influence object is Thing (ie. anything).

Here we introduce two special purpose association classes that are used
to constraint the ranges of object and roles. This effectively
implements a two-step selection process. For any `influenced_by`, a UI
user can for select the approximate type of influence, and only in step
2 detail it.

The hope is that this can better account for the real-world complexity
of annotation systems.