Remove graph-oriented schemas
The development of the current FLAT schema setup has made them obsolete.
Combined with the advances of the shacl-vue editor, it is now possible
to enter fairly complex qualified relationships, even via a GUI. The
expressiveness of this approach effectively surpasses that of the
schemas that are being removed here. Hence there is no reason to keep
them.