Make `part_of` generally `multivalued: true`
A mismatch between two derived classes (that also added a mixin) brought
LinkML to its knees. This made me look through all usage of `part_of`.

It is now made `multivalued: true` everywhere, except for the empirical
data demo (for historical reasons).

In general it makes sense to have that in all the present use cases. I
could not convince myself that this is true in any and all cases, so I
kept the default `false` for now in the slot definition.