Establish `FlatProperty` with the `FlatThing` curation/UI slots
All other changes are more-or-less consequences.

This is done, because before it was impossible to name properties
(custom slots). `things` doesn't have `name`, and `Property` did not
have `display_label` either.

This change also prepared for https://github.com/psychoinformatics-de/datalad-concepts/issues/479
which proposed to merge `flat` into `things` and make the result the
basis for `things/v2`